backend/requirement.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
bcrypt
passlib
email-validator